Fix github's bullshit
It seems the "ubuntu latest" image is rather old and has an old
list of packages, so it needs to run an apt update to get the new
list of packages.

Otherwise apt will try to download older versions that are no longer on
the apt mirror and fail with a 404.

If the image was truly recent this wouldn't be needed or it'd just be a
temporary issue.